fix: rework backup
